requirements.txt
django==3.2.9
django-simpleui==2021.11.5
pandas==1.3.4
pyecharts==1.9.1
django_notifications_hq=1.6.0
pillow==9.0.1
python-docx==0.8.11
lxml==4.8.0
docxtpl==0.15.2
docxcompose==1.3.4
pydocx==0.9.10
pycryptodome==3.14.1